+++ core/src/main/java/org/patternfly/core/Attributes.java
@@ -21,5 +21,6 @@ public interface Attributes {
     String tabindex = "tabindex";
     String role = "role";
     String required = "required";
+    String scope = "scope";
     String type = "type";
 }