+++ CHANGELOG.md
@@ -6,6 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add single and multi-selection handlers to the table component.
+- Add `MenuItem.menuItem(String identifier)` factory method.
+- Add `Button.button(PredefinedIcon icon, String href)`
+- Add `Button.button(PredefinedIcon icon, String href, String target)`
+
+### Fixed
+
+- Fix `Checkbox.standalone(boolean)`
+- Fix `Tbody.empty()`
+- Fix NPE when using `Ordered.defaultOrder()`
+- Fix `PageMain.clear()` method and `Elements.removeChildrenFrom()` usage in `PageMain`.
+
+### Upgrades
+
+- Bump JBoss Parent to 52
+
 ## [0.4.17] - 2026-02-12
 
 ### Added